refactoring shared methods
