+++ acpreprocessing/stitching_modules/convert_to_n5/ims_to_ngff.py
@@ -19,582 +19,582 @@ import argschema
 
 import acpreprocessing.utils.convert
 import acpreprocessing.stitching_modules.convert_to_n5.psdeskew as psd
-
-
-def iterate_chunks(ds, slice_length, interleaved_channels=1, channel=0, interleaving_offset=0):
-    """given an iterator, iterate over tuples of a
-    given length from that iterator
-
-    Parameters
-    ----------
-    it : iterable
-        iterable object to iterate over in slices of slice_length
-    slice_length : int
-        number of items in iterator per chunk
-
-    Yields
-    ------
-    chunk : tuple
-        tuple of items from iterator of size slice_length. Yields
-        remainder of iterator when exhausted
-    """
-    start = 0
-    stop = slice_length
-    chunk = ds[start:stop]
-    while start<ds.shape[0]:
-        yield chunk
-        start = stop
-        stop = start + slice_length
-        if stop < ds.shape[0]:
-            chunk = ds[start:stop]
-        else:
-            chunk = ds[start:]
-
-
-def iter_arrays(r, interleaved_channels=1, channel=0, interleaving_offset=0):
-    """iterate arrays from an imageio tiff reader.  Allows the last image
-    of the array to be None, as is the case for data with 'dropped frames'.
-
-    Parameters
-    ----------
-    r : imageio.reader
-        imageio tiff reader from which to iterate multitiff arrays
-    interleaved_channels : int
-        number of channels interleaved in tiff when selecting which
-        arrays to read
-    channel : int
-        channel to iterate over in case of interleaved tiffs
-
-    Yields
-    -------
-    arr : numpy.ndarray
-        constituent page array of reader r
-    """
-    for i, p in enumerate(r._tf.pages):
-        page_channel = (i + interleaving_offset) % interleaved_channels
-        if page_channel != channel:
-            continue
-        arr = p.asarray()
-        if arr is not None:
-            yield arr
-        elif i == r.get_length() - 1:
-            continue
-        else:
-            raise ValueError
-
-
-def iterate_2d_arrays_from_mimgfns(mimgfns, interleaved_channels=1, channel=0):
-    """iterate constituent arrays from an iterator of image filenames
-    that can be opened as an imageio multi-image.
-
-    Parameters
-    ----------
-    mimgfns : list of str
-        imageio-compatible name inputs to be opened as multi-images
-
-    Yields
-    ------
-    array : numpy.ndarray
-        constituent array from ordered iterator of multi-image files
-    """
-    offset = 0
-    for mimgfn in mimgfns:
-        with imageio.get_reader(mimgfn, mode="I") as r:
-            yield from iter_arrays(r, interleaved_channels, channel, offset)
-            offset = (offset + r.get_length()) % interleaved_channels
-
-
-def iterate_numpy_chunks_from_dataset(
-        dataset, slice_length=None, pad=True, *args, **kwargs):
-    """iterate over a contiguous hdf5 daataset as chunks of numpy arrays
-
-    Parameters
-    ----------
-    dataset : hdf5 dataset
-        imageio-compatible name inputs to be opened as multi-images
-    slice_length : int
-        number of 2d arrays from mimgfns included per chunk
-    pad : bool, optional
-        whether to extend final returned chunks with zeros
-
-    Yields
-    ------
-    arr : numpy.ndarray
-        3D numpy array representing a consecutive chunk of 2D arrays
-    """
-    #array_gen = iterate_2d_arrays_from_dataset(mimgfns, *args, **kwargs)
-    for chunk in iterate_chunks(dataset, slice_length,*args,**kwargs):
-        arr = numpy.asarray(chunk)
-        if pad:
-            if arr.shape[0] != slice_length:
-                newarr = numpy.zeros((slice_length, *arr.shape[1:]),
-                                     dtype=arr.dtype)
-                newarr[:arr.shape[0], :, :] = arr[:, :, :]
-                yield newarr
-            else:
-                yield arr
-        else:
-            yield arr
-
-
-def length_to_interleaved_length(length, interleaved_channels):
-    """get length of each interleaved channel that contributes to a
-    total array length
-
-    Parameters
-    ----------
-    length : int
-        length of interleaved array
-    interleaved_channels : int
-        number of constituent channels in interleaved array
-
-    Returns
-    -------
-    channel_lengths : list of int
-        length of each constituent channel in interleaved array
-    """
-    interleaved_length, r = divmod(length, interleaved_channels)
-    return [
-        interleaved_length + (1 if ic < r else 0)
-        for ic in range(interleaved_channels)
-    ]
-
-
-def mimg_shape_from_fn(mimg_fn, only_length_tup=False):
-    """get the shape of an imageio multi-image file without
-    reading it as a volume
-
-    Parameters
-    ----------
-    mimg_fn : str
-        imageio-compatible name input to be opened as multi-image
-    only_length_tup : bool, optional
-        whether to return only the number of frames or
-        include the dimensions of a representative frame
-
-    Returns
-    -------
-    shape : tuple of int
-        shape of array defined by mimg_fn
-    """
-    with imageio.get_reader(mimg_fn, mode="I") as r:
-        l = r.get_length()
-
-        if only_length_tup:
-            s = (l,)
-        else:
-            s = (l, *r.get_data(0).shape)
-    return s
-
-
-def array_downsample_chunks(arr, ds_factors, block_shape, **kwargs):
-    """downsample a numpy array by inputting subvolumes
-    to acpreprocessing.utils.convert.downsample_stack_volume
-
-    Parameters
-    ----------
-    arr : numpy.ndarray
-        array to downsample
-    ds_factors : tuple of int
-        integer downsample factor by which arr should be downsamples.
-        Length must match arr.ndim
-    block_shape : tuple of int
-        shape of individual blocks to process
-
-    Returns
-    -------
-    output_arr : numpy.ndarray
-        downsampled version of arr
-    """
-    blocks = skimage.util.view_as_blocks(arr, block_shape)
-
-    # construct output array and blocks
-    output_shape = tuple(
-        int(math.ceil(s / f)) for s, f in zip(arr.shape, ds_factors))
-    output_block_shape = tuple(
-        int(math.ceil(s / f)) for s, f in zip(block_shape, ds_factors))
-    output_arr = numpy.empty(output_shape, dtype=arr.dtype)
-    output_blocks = skimage.util.view_as_blocks(
-        output_arr, output_block_shape)
-
-    # process
-    for idxs in numpy.ndindex(blocks.shape[:arr.ndim]):
-        output_blocks[idxs] = (
-            acpreprocessing.utils.convert.downsample_stack_volume(
-                blocks[idxs], ds_factors, **kwargs))
-
-    return output_arr
-
-
-def array_downsample_threaded(arr, ds_factors, block_shape,
-                              n_threads=3, **kwargs):
-    """downsample a numpy array by concurrently inputting subvolumes
-    to acpreprocessing.utils.convert.downsample_stack_volume
-
-    Parameters
-    ----------
-    arr : numpy.ndarray
-        array to downsample
-    ds_factors : tuple of int
-        integer downsample factor by which arr should be downsamples.
-        Length must match arr.ndim
-    block_shape : tuple of int
-        shape of individual blocks to process
-    n_threads :  int
-        number of threads used to concurrently downsample chunks
-
-    Returns
-    -------
-    output_arr : numpy.ndarray
-        downsampled version of arr
-    """
-    blocks = skimage.util.view_as_blocks(arr, block_shape)
-
-    # construct output array and blocks
-    output_shape = tuple(
-        int(math.ceil(s / f)) for s, f in zip(arr.shape, ds_factors))
-    output_block_shape = tuple(
-        int(math.ceil(s / f)) for s, f in zip(block_shape, ds_factors))
-    output_arr = numpy.empty(output_shape, dtype=arr.dtype)
-    output_blocks = skimage.util.view_as_blocks(
-        output_arr, output_block_shape)
-
-    # process
-    with concurrent.futures.ThreadPoolExecutor(max_workers=n_threads) as e:
-        fut_to_idxs = {}
-        for idxs in numpy.ndindex(blocks.shape[:arr.ndim]):
-            fut_to_idxs[e.submit(
-                acpreprocessing.utils.convert.downsample_stack_volume,
-                blocks[idxs], ds_factors, **kwargs)] = idxs
-
-        for fut in concurrent.futures.as_completed(fut_to_idxs.keys()):
-            idxs = fut_to_idxs[fut]
-            output_blocks[idxs] = fut.result()
-    return output_arr
-
-
-class ArrayChunkingError(ValueError):
-    """ValueError caused by incorrect array chunking"""
-
-
-def downsample_array(arr, *args, n_threads=None, block_shape=None,
-                     block_divs=None, allow_failover=True, **kwargs):
-    """downsample array with optional subvolume chunking and threading
-
-    Parameters
-    ----------
-    arr : numpy.ndarray
-        array to downsample
-    ds_factors : tuple of int
-        integer downsample factor by which arr should be downsamples.
-        Length must match arr.ndim
-    block_shape : tuple of int
-        shape of individual blocks to process.
-        Supersedes block_divs if both are provided
-    block_divs : tuple of int
-        number of divisions for array chunking.
-    n_threads :  int
-        number of threads used to concurrently downsample chunks
-    allow_failover : bool, optional
-        whether to default to using
-        acpreprocessing.utils.convert.downsample_stack_volume when
-        block_shape or block_divs is not actionable.  (default: True)
-
-    Returns
-    -------
-    output_arr : numpy.ndarray
-        downsampled version of arr
-    """
-    try:
-        if block_shape is not None or block_divs is not None:
-            if block_shape is None:
-                block_shape = [a_s / d for a_s, d in zip(
-                    arr.shape, block_divs)]
-                if not all([s.is_integer() for s in block_shape]):
-                    raise ArrayChunkingError(
-                        "cannot divide array of shape {} by {}".format(
-                            arr.shape, block_divs))
-                block_shape = tuple(int(s) for s in block_shape)
-            if n_threads is not None:
-                return array_downsample_threaded(
-                    arr, *args, n_threads=n_threads,
-                    block_shape=block_shape, **kwargs)
-            return array_downsample_chunks(
-                arr, *args, block_shape=block_shape, **kwargs)
-    except ArrayChunkingError:
-        if not allow_failover:
-            raise
-    return acpreprocessing.utils.convert.downsample_stack_volume(
-        arr, *args, **kwargs)
-
-
-def mip_level_shape(lvl, lvl_0_shape):
-    """get the expected shape of a MIP level downsample for a given
-    full resolution shape.  Expects 2x downsampling on each dimension.
-
-    Parameters
-    ----------
-    lvl : int
-        mip level for which shape is calculated
-    lvl_0_shape : tuple of int
-        shape tuple of level 0
-
-    Returns
-    -------
-    lvl_shape : tuple of int
-        shape tuple for lvl
-    """
-    return tuple([*map(lambda x: math.ceil(x / (2**lvl)), lvl_0_shape)])
-
-
-def dswrite_chunk(ds, start, end, arr, silent_overflow=True):
-    """write array arr into array-like n5 dataset defined by
-    start and end points on 0 axis.
-
-    Parameters
-    ----------
-    ds : z5py.dataset.Dataset (n5 3D) or zarr.dataset (zarr 5D)
-        array-like dataset to fill in
-    start : int
-        start index along 0 axis of ds to fill in
-    end : int
-        end index along 0 axis of ds to fill in
-    arr : numpy.ndarray
-        array from which ds values will be taken
-    silent_overflow : bool, optional
-        whether to shrink the end index to match the
-        shape of ds (default: True)
-    """
-    if len(ds.shape) == 5:  # dataset dimensions should be 3 or 5
-        if end >= ds.shape[2] and silent_overflow:
-            end = ds.shape[2]
-        if end > start:
-            ds[0, 0, start:end, :, :] = arr[:(end - start), :, :]
-    elif len(ds.shape) == 3:
-        if end >= ds.shape[0] and silent_overflow:
-            end = ds.shape[0]
-        if end > start:
-            ds[start:end, :, :] = arr[:(end - start), :, :]
-
-
-@dataclasses.dataclass
-class MIPArray:
-    lvl: int
-    array: numpy.ndarray
-    start: int
-    end: int
-
-
-def iterate_mip_levels_from_dataset(
-        dataset, lvl, block_size, slice_length, downsample_factor,
-        downsample_method=None, lvl_to_mip_kwargs=None,
-        interleaved_channels=1, channel=0, deskew_kwargs={}):
-    """recursively generate MIPmap levels from an iterator of multi-image files
-
-    Parameters
-    ----------
-    dataset : hdf5 dataset
-        imageio-compatible name inputs to be opened as multi-images
-    lvl : int
-        integer mip level to generate
-    block_size : int
-        number of 2D arrays in chunk to process for a chunk at lvl
-    slice_length : int
-        number of 2D arrays to gather from tiff stacks
-    downsample_factor : tuple of int
-        integer downsampling factor for MIP levels
-    downsample_method : str, optional
-        downsampling method for
-        acpreprocessing.utils.convert.downsample_stack_volume
-    lvl_to_mip_kwargs :  dict, optional
-        mapping of MIP level to kwargs used in MIPmap generation
-    interleaved_channels : int
-        number of channels interleaved in the tiff files (default 1)
-    channel : int, optional
-        channel from which interleaved data should be read (default 0)
-    deskew_kwargs : dict, optional
-        parameters for pixel shifting deskew
-
-    Yields
-    ------
-    ma : acpreprocessing.stitching_modules.convert_to_n5.tiff_to_n5.MipArray
-        object describing chunked array, MIP level of origin, and chunk indices
-    """
-    lvl_to_mip_kwargs = ({} if lvl_to_mip_kwargs is None
-                         else lvl_to_mip_kwargs)
-    mip_kwargs = lvl_to_mip_kwargs.get(lvl, {})
-    start_index = 0
-    chunk_index = 0
-    if lvl > 0:
-        num_chunks = downsample_factor[0]
-
-        i = 0
-        for ma in iterate_mip_levels_from_dataset(
-                dataset, lvl-1, block_size, slice_length,
-                downsample_factor, downsample_method,
-                lvl_to_mip_kwargs, interleaved_channels=interleaved_channels,
-                channel=channel, deskew_kwargs=deskew_kwargs):
-            chunk = ma.array
-            # throw array up for further processing
-            yield ma
-            # ignore if not parent resolution
-            if ma.lvl != lvl-1:
-                continue
-
-            try:
-                temp_lminus1_arr
-            except NameError:
-                temp_lminus1_arr = numpy.empty(
-                    (num_chunks*chunk.shape[0], *chunk.shape[1:]),
-                    dtype=chunk.dtype)
-
-            # fill in temporary block according to index
-            chunk_size = chunk.shape[0]
-            block_offset = i*block_size
-
-            temp_lminus1_arr[
-                block_offset:block_offset+chunk_size, :, :] = chunk[:, :, :]
-
-            # copy op only for uneven final chunk
-            if chunk_size != block_size:
-                temp_lminus1_arr = temp_lminus1_arr[
-                    :block_offset+chunk_size, :, :]
-
-            if i == num_chunks - 1:
-                temp_arr = (
-                    downsample_array(
-                        temp_lminus1_arr, downsample_factor, dtype=chunk.dtype,
-                        method=downsample_method, **mip_kwargs))
-                end_index = start_index + temp_arr.shape[0]
-                yield MIPArray(lvl, temp_arr, start_index, end_index)
-                start_index += temp_arr.shape[0]
-                i = 0
-            else:
-                i += 1
-        # add any leftovers
-        if i > 0:
-            temp_arr = acpreprocessing.utils.convert.downsample_stack_volume(
-                temp_lminus1_arr, downsample_factor, dtype=chunk.dtype,
-                method=downsample_method)
-            end_index = start_index + temp_arr.shape[0]
-            yield MIPArray(lvl, temp_arr, start_index, end_index)
-    else:
-        # get level 0 chunks
-        # block_size is the number of slices to read from tiffs
-        for chunk in iterate_numpy_chunks_from_dataset(
-                dataset, slice_length, pad=False,
-                interleaved_channels=interleaved_channels,
-                channel=channel):
-            # deskew level 0 chunk
-            if deskew_kwargs:
-                chunk = numpy.transpose(psd.deskew_block(
-                    chunk, chunk_index, **deskew_kwargs), (2, 1, 0))
-            end_index = start_index + chunk.shape[0]
-            yield MIPArray(lvl, chunk, start_index, end_index)
-            start_index += chunk.shape[0]
-            chunk_index += 1
-
-
-
-def omezarr_attrs(name, position_xyz, lvl0_xyz_res, max_lvl):
-    # TODO only single channel implemented
-    datasets = []
-    for i in range(max_lvl+1):
-        s = [(2**i)*r for r in lvl0_xyz_res]
-        d = {"coordinateTransformations": [
-            {"scale": [
-                1.0,
-                1.0,
-                s[2],
-                s[1],
-                s[0]
-            ],
-                "type": "scale"
-            }
-        ],
-            "path": str(i)
-        }
-        datasets.append(d)
-    attrs = {}
-    attrs["multiscales"] = [
-        {
-            "axes": [
-                {
-                    "name": "t",
-                    "type": "time",
-                    "unit": "millisecond"
-                },
-                {
-                    "name": "c",
-                    "type": "channel"
-                },
-                {
-                    "name": "z",
-                    "type": "space",
-                    "unit": "micrometer"
-                },
-                {
-                    "name": "y",
-                    "type": "space",
-                    "unit": "micrometer"
-                },
-                {
-                    "name": "x",
-                    "type": "space",
-                    "unit": "micrometer"
-                }
-            ],
-            "datasets": datasets,
-            "name": "/" + name,
-            "translation": [
-                0,
-                0,
-                0.0,
-                0.0,
-                0.0
-            ],
-            "version": "0.4",
-            "coordinateTransformations": [
-                {
-                    "translation": [
-                        0.0,
-                        0.0,
-                        position_xyz[2],
-                        position_xyz[1],
-                        position_xyz[0]
-                    ],
-                    "type": "translation"
-                }
-            ]
-        }
-    ]
-    attrs["omero"] = {
-        "channels": [
-            {
-                "active": True,
-                "coefficient": 1,
-                "color": "000000",
-                "family": "linear",
-                "inverted": False,
-                "label": "Channel:" + name + ":0",
-                "window": {
-                    "end": 1.0,
-                    "max": 1.0,
-                    "min": 0.0,
-                    "start": 0.0
-                }
-            }
-        ],
-        "id": 1,
-        "name": name,
-        "rdefs": {
-            "defaultT": 0,
-            "defaultZ": 288,
-            "model": "color"
-        },
-        "version": "0.4"
-    }
-    return attrs
+from acpreprocessing.stitching_modules.convert_to_n5.tiff_to_ngff import iterate_mip_levels_from_dataset,mip_level_shape,dswrite_chunk,omezarr_attrs,TiffDirToZarrInputParameters
+
+# def iterate_chunks(ds, slice_length, interleaved_channels=1, channel=0, interleaving_offset=0):
+#     """given an iterator, iterate over tuples of a
+#     given length from that iterator
+
+#     Parameters
+#     ----------
+#     it : iterable
+#         iterable object to iterate over in slices of slice_length
+#     slice_length : int
+#         number of items in iterator per chunk
+
+#     Yields
+#     ------
+#     chunk : tuple
+#         tuple of items from iterator of size slice_length. Yields
+#         remainder of iterator when exhausted
+#     """
+#     start = 0
+#     stop = slice_length
+#     chunk = ds[start:stop]
+#     while start<ds.shape[0]:
+#         yield chunk
+#         start = stop
+#         stop = start + slice_length
+#         if stop < ds.shape[0]:
+#             chunk = ds[start:stop]
+#         else:
+#             chunk = ds[start:]
+
+
+# def iter_arrays(r, interleaved_channels=1, channel=0, interleaving_offset=0):
+#     """iterate arrays from an imageio tiff reader.  Allows the last image
+#     of the array to be None, as is the case for data with 'dropped frames'.
+
+#     Parameters
+#     ----------
+#     r : imageio.reader
+#         imageio tiff reader from which to iterate multitiff arrays
+#     interleaved_channels : int
+#         number of channels interleaved in tiff when selecting which
+#         arrays to read
+#     channel : int
+#         channel to iterate over in case of interleaved tiffs
+
+#     Yields
+#     -------
+#     arr : numpy.ndarray
+#         constituent page array of reader r
+#     """
+#     for i, p in enumerate(r._tf.pages):
+#         page_channel = (i + interleaving_offset) % interleaved_channels
+#         if page_channel != channel:
+#             continue
+#         arr = p.asarray()
+#         if arr is not None:
+#             yield arr
+#         elif i == r.get_length() - 1:
+#             continue
+#         else:
+#             raise ValueError
+
+
+# def iterate_2d_arrays_from_mimgfns(mimgfns, interleaved_channels=1, channel=0):
+#     """iterate constituent arrays from an iterator of image filenames
+#     that can be opened as an imageio multi-image.
+
+#     Parameters
+#     ----------
+#     mimgfns : list of str
+#         imageio-compatible name inputs to be opened as multi-images
+
+#     Yields
+#     ------
+#     array : numpy.ndarray
+#         constituent array from ordered iterator of multi-image files
+#     """
+#     offset = 0
+#     for mimgfn in mimgfns:
+#         with imageio.get_reader(mimgfn, mode="I") as r:
+#             yield from iter_arrays(r, interleaved_channels, channel, offset)
+#             offset = (offset + r.get_length()) % interleaved_channels
+
+
+# def iterate_numpy_chunks_from_dataset(
+#         dataset, slice_length=None, pad=True, *args, **kwargs):
+#     """iterate over a contiguous hdf5 daataset as chunks of numpy arrays
+
+#     Parameters
+#     ----------
+#     dataset : hdf5 dataset
+#         imageio-compatible name inputs to be opened as multi-images
+#     slice_length : int
+#         number of 2d arrays from mimgfns included per chunk
+#     pad : bool, optional
+#         whether to extend final returned chunks with zeros
+
+#     Yields
+#     ------
+#     arr : numpy.ndarray
+#         3D numpy array representing a consecutive chunk of 2D arrays
+#     """
+#     #array_gen = iterate_2d_arrays_from_dataset(mimgfns, *args, **kwargs)
+#     for chunk in iterate_chunks(dataset, slice_length,*args,**kwargs):
+#         arr = numpy.asarray(chunk)
+#         if pad:
+#             if arr.shape[0] != slice_length:
+#                 newarr = numpy.zeros((slice_length, *arr.shape[1:]),
+#                                       dtype=arr.dtype)
+#                 newarr[:arr.shape[0], :, :] = arr[:, :, :]
+#                 yield newarr
+#             else:
+#                 yield arr
+#         else:
+#             yield arr
+
+
+# def length_to_interleaved_length(length, interleaved_channels):
+#     """get length of each interleaved channel that contributes to a
+#     total array length
+
+#     Parameters
+#     ----------
+#     length : int
+#         length of interleaved array
+#     interleaved_channels : int
+#         number of constituent channels in interleaved array
+
+#     Returns
+#     -------
+#     channel_lengths : list of int
+#         length of each constituent channel in interleaved array
+#     """
+#     interleaved_length, r = divmod(length, interleaved_channels)
+#     return [
+#         interleaved_length + (1 if ic < r else 0)
+#         for ic in range(interleaved_channels)
+#     ]
+
+
+# def mimg_shape_from_fn(mimg_fn, only_length_tup=False):
+#     """get the shape of an imageio multi-image file without
+#     reading it as a volume
+
+#     Parameters
+#     ----------
+#     mimg_fn : str
+#         imageio-compatible name input to be opened as multi-image
+#     only_length_tup : bool, optional
+#         whether to return only the number of frames or
+#         include the dimensions of a representative frame
+
+#     Returns
+#     -------
+#     shape : tuple of int
+#         shape of array defined by mimg_fn
+#     """
+#     with imageio.get_reader(mimg_fn, mode="I") as r:
+#         l = r.get_length()
+
+#         if only_length_tup:
+#             s = (l,)
+#         else:
+#             s = (l, *r.get_data(0).shape)
+#     return s
+
+
+# def array_downsample_chunks(arr, ds_factors, block_shape, **kwargs):
+#     """downsample a numpy array by inputting subvolumes
+#     to acpreprocessing.utils.convert.downsample_stack_volume
+
+#     Parameters
+#     ----------
+#     arr : numpy.ndarray
+#         array to downsample
+#     ds_factors : tuple of int
+#         integer downsample factor by which arr should be downsamples.
+#         Length must match arr.ndim
+#     block_shape : tuple of int
+#         shape of individual blocks to process
+
+#     Returns
+#     -------
+#     output_arr : numpy.ndarray
+#         downsampled version of arr
+#     """
+#     blocks = skimage.util.view_as_blocks(arr, block_shape)
+
+#     # construct output array and blocks
+#     output_shape = tuple(
+#         int(math.ceil(s / f)) for s, f in zip(arr.shape, ds_factors))
+#     output_block_shape = tuple(
+#         int(math.ceil(s / f)) for s, f in zip(block_shape, ds_factors))
+#     output_arr = numpy.empty(output_shape, dtype=arr.dtype)
+#     output_blocks = skimage.util.view_as_blocks(
+#         output_arr, output_block_shape)
+
+#     # process
+#     for idxs in numpy.ndindex(blocks.shape[:arr.ndim]):
+#         output_blocks[idxs] = (
+#             acpreprocessing.utils.convert.downsample_stack_volume(
+#                 blocks[idxs], ds_factors, **kwargs))
+
+#     return output_arr
+
+
+# def array_downsample_threaded(arr, ds_factors, block_shape,
+#                               n_threads=3, **kwargs):
+#     """downsample a numpy array by concurrently inputting subvolumes
+#     to acpreprocessing.utils.convert.downsample_stack_volume
+
+#     Parameters
+#     ----------
+#     arr : numpy.ndarray
+#         array to downsample
+#     ds_factors : tuple of int
+#         integer downsample factor by which arr should be downsamples.
+#         Length must match arr.ndim
+#     block_shape : tuple of int
+#         shape of individual blocks to process
+#     n_threads :  int
+#         number of threads used to concurrently downsample chunks
+
+#     Returns
+#     -------
+#     output_arr : numpy.ndarray
+#         downsampled version of arr
+#     """
+#     blocks = skimage.util.view_as_blocks(arr, block_shape)
+
+#     # construct output array and blocks
+#     output_shape = tuple(
+#         int(math.ceil(s / f)) for s, f in zip(arr.shape, ds_factors))
+#     output_block_shape = tuple(
+#         int(math.ceil(s / f)) for s, f in zip(block_shape, ds_factors))
+#     output_arr = numpy.empty(output_shape, dtype=arr.dtype)
+#     output_blocks = skimage.util.view_as_blocks(
+#         output_arr, output_block_shape)
+
+#     # process
+#     with concurrent.futures.ThreadPoolExecutor(max_workers=n_threads) as e:
+#         fut_to_idxs = {}
+#         for idxs in numpy.ndindex(blocks.shape[:arr.ndim]):
+#             fut_to_idxs[e.submit(
+#                 acpreprocessing.utils.convert.downsample_stack_volume,
+#                 blocks[idxs], ds_factors, **kwargs)] = idxs
+
+#         for fut in concurrent.futures.as_completed(fut_to_idxs.keys()):
+#             idxs = fut_to_idxs[fut]
+#             output_blocks[idxs] = fut.result()
+#     return output_arr
+
+
+# class ArrayChunkingError(ValueError):
+#     """ValueError caused by incorrect array chunking"""
+
+
+# def downsample_array(arr, *args, n_threads=None, block_shape=None,
+#                       block_divs=None, allow_failover=True, **kwargs):
+#     """downsample array with optional subvolume chunking and threading
+
+#     Parameters
+#     ----------
+#     arr : numpy.ndarray
+#         array to downsample
+#     ds_factors : tuple of int
+#         integer downsample factor by which arr should be downsamples.
+#         Length must match arr.ndim
+#     block_shape : tuple of int
+#         shape of individual blocks to process.
+#         Supersedes block_divs if both are provided
+#     block_divs : tuple of int
+#         number of divisions for array chunking.
+#     n_threads :  int
+#         number of threads used to concurrently downsample chunks
+#     allow_failover : bool, optional
+#         whether to default to using
+#         acpreprocessing.utils.convert.downsample_stack_volume when
+#         block_shape or block_divs is not actionable.  (default: True)
+
+#     Returns
+#     -------
+#     output_arr : numpy.ndarray
+#         downsampled version of arr
+#     """
+#     try:
+#         if block_shape is not None or block_divs is not None:
+#             if block_shape is None:
+#                 block_shape = [a_s / d for a_s, d in zip(
+#                     arr.shape, block_divs)]
+#                 if not all([s.is_integer() for s in block_shape]):
+#                     raise ArrayChunkingError(
+#                         "cannot divide array of shape {} by {}".format(
+#                             arr.shape, block_divs))
+#                 block_shape = tuple(int(s) for s in block_shape)
+#             if n_threads is not None:
+#                 return array_downsample_threaded(
+#                     arr, *args, n_threads=n_threads,
+#                     block_shape=block_shape, **kwargs)
+#             return array_downsample_chunks(
+#                 arr, *args, block_shape=block_shape, **kwargs)
+#     except ArrayChunkingError:
+#         if not allow_failover:
+#             raise
+#     return acpreprocessing.utils.convert.downsample_stack_volume(
+#         arr, *args, **kwargs)
+
+
+# def mip_level_shape(lvl, lvl_0_shape):
+#     """get the expected shape of a MIP level downsample for a given
+#     full resolution shape.  Expects 2x downsampling on each dimension.
+
+#     Parameters
+#     ----------
+#     lvl : int
+#         mip level for which shape is calculated
+#     lvl_0_shape : tuple of int
+#         shape tuple of level 0
+
+#     Returns
+#     -------
+#     lvl_shape : tuple of int
+#         shape tuple for lvl
+#     """
+#     return tuple([*map(lambda x: math.ceil(x / (2**lvl)), lvl_0_shape)])
+
+
+# def dswrite_chunk(ds, start, end, arr, silent_overflow=True):
+#     """write array arr into array-like n5 dataset defined by
+#     start and end points on 0 axis.
+
+#     Parameters
+#     ----------
+#     ds : z5py.dataset.Dataset (n5 3D) or zarr.dataset (zarr 5D)
+#         array-like dataset to fill in
+#     start : int
+#         start index along 0 axis of ds to fill in
+#     end : int
+#         end index along 0 axis of ds to fill in
+#     arr : numpy.ndarray
+#         array from which ds values will be taken
+#     silent_overflow : bool, optional
+#         whether to shrink the end index to match the
+#         shape of ds (default: True)
+#     """
+#     if len(ds.shape) == 5:  # dataset dimensions should be 3 or 5
+#         if end >= ds.shape[2] and silent_overflow:
+#             end = ds.shape[2]
+#         if end > start:
+#             ds[0, 0, start:end, :, :] = arr[:(end - start), :, :]
+#     elif len(ds.shape) == 3:
+#         if end >= ds.shape[0] and silent_overflow:
+#             end = ds.shape[0]
+#         if end > start:
+#             ds[start:end, :, :] = arr[:(end - start), :, :]
+
+
+# @dataclasses.dataclass
+# class MIPArray:
+#     lvl: int
+#     array: numpy.ndarray
+#     start: int
+#     end: int
+
+
+# def iterate_mip_levels_from_dataset(
+#         dataset, lvl, block_size, slice_length, downsample_factor,
+#         downsample_method=None, lvl_to_mip_kwargs=None,
+#         interleaved_channels=1, channel=0, deskew_kwargs={}):
+#     """recursively generate MIPmap levels from an iterator of multi-image files
+
+#     Parameters
+#     ----------
+#     dataset : hdf5 dataset
+#         imageio-compatible name inputs to be opened as multi-images
+#     lvl : int
+#         integer mip level to generate
+#     block_size : int
+#         number of 2D arrays in chunk to process for a chunk at lvl
+#     slice_length : int
+#         number of 2D arrays to gather from tiff stacks
+#     downsample_factor : tuple of int
+#         integer downsampling factor for MIP levels
+#     downsample_method : str, optional
+#         downsampling method for
+#         acpreprocessing.utils.convert.downsample_stack_volume
+#     lvl_to_mip_kwargs :  dict, optional
+#         mapping of MIP level to kwargs used in MIPmap generation
+#     interleaved_channels : int
+#         number of channels interleaved in the tiff files (default 1)
+#     channel : int, optional
+#         channel from which interleaved data should be read (default 0)
+#     deskew_kwargs : dict, optional
+#         parameters for pixel shifting deskew
+
+#     Yields
+#     ------
+#     ma : acpreprocessing.stitching_modules.convert_to_n5.tiff_to_n5.MipArray
+#         object describing chunked array, MIP level of origin, and chunk indices
+#     """
+#     lvl_to_mip_kwargs = ({} if lvl_to_mip_kwargs is None
+#                           else lvl_to_mip_kwargs)
+#     mip_kwargs = lvl_to_mip_kwargs.get(lvl, {})
+#     start_index = 0
+#     chunk_index = 0
+#     if lvl > 0:
+#         num_chunks = downsample_factor[0]
+
+#         i = 0
+#         for ma in iterate_mip_levels_from_dataset(
+#                 dataset, lvl-1, block_size, slice_length,
+#                 downsample_factor, downsample_method,
+#                 lvl_to_mip_kwargs, interleaved_channels=interleaved_channels,
+#                 channel=channel, deskew_kwargs=deskew_kwargs):
+#             chunk = ma.array
+#             # throw array up for further processing
+#             yield ma
+#             # ignore if not parent resolution
+#             if ma.lvl != lvl-1:
+#                 continue
+
+#             try:
+#                 temp_lminus1_arr
+#             except NameError:
+#                 temp_lminus1_arr = numpy.empty(
+#                     (num_chunks*chunk.shape[0], *chunk.shape[1:]),
+#                     dtype=chunk.dtype)
+
+#             # fill in temporary block according to index
+#             chunk_size = chunk.shape[0]
+#             block_offset = i*block_size
+
+#             temp_lminus1_arr[
+#                 block_offset:block_offset+chunk_size, :, :] = chunk[:, :, :]
+
+#             # copy op only for uneven final chunk
+#             if chunk_size != block_size:
+#                 temp_lminus1_arr = temp_lminus1_arr[
+#                     :block_offset+chunk_size, :, :]
+
+#             if i == num_chunks - 1:
+#                 temp_arr = (
+#                     downsample_array(
+#                         temp_lminus1_arr, downsample_factor, dtype=chunk.dtype,
+#                         method=downsample_method, **mip_kwargs))
+#                 end_index = start_index + temp_arr.shape[0]
+#                 yield MIPArray(lvl, temp_arr, start_index, end_index)
+#                 start_index += temp_arr.shape[0]
+#                 i = 0
+#             else:
+#                 i += 1
+#         # add any leftovers
+#         if i > 0:
+#             temp_arr = acpreprocessing.utils.convert.downsample_stack_volume(
+#                 temp_lminus1_arr, downsample_factor, dtype=chunk.dtype,
+#                 method=downsample_method)
+#             end_index = start_index + temp_arr.shape[0]
+#             yield MIPArray(lvl, temp_arr, start_index, end_index)
+#     else:
+#         # get level 0 chunks
+#         # block_size is the number of slices to read from tiffs
+#         for chunk in iterate_numpy_chunks_from_dataset(
+#                 dataset, slice_length, pad=False,
+#                 interleaved_channels=interleaved_channels,
+#                 channel=channel):
+#             # deskew level 0 chunk
+#             if deskew_kwargs:
+#                 chunk = numpy.transpose(psd.deskew_block(
+#                     chunk, chunk_index, **deskew_kwargs), (2, 1, 0))
+#             end_index = start_index + chunk.shape[0]
+#             yield MIPArray(lvl, chunk, start_index, end_index)
+#             start_index += chunk.shape[0]
+#             chunk_index += 1
+
+
+
+# def omezarr_attrs(name, position_xyz, lvl0_xyz_res, max_lvl):
+#     # TODO only single channel implemented
+#     datasets = []
+#     for i in range(max_lvl+1):
+#         s = [(2**i)*r for r in lvl0_xyz_res]
+#         d = {"coordinateTransformations": [
+#             {"scale": [
+#                 1.0,
+#                 1.0,
+#                 s[2],
+#                 s[1],
+#                 s[0]
+#             ],
+#                 "type": "scale"
+#             }
+#         ],
+#             "path": str(i)
+#         }
+#         datasets.append(d)
+#     attrs = {}
+#     attrs["multiscales"] = [
+#         {
+#             "axes": [
+#                 {
+#                     "name": "t",
+#                     "type": "time",
+#                     "unit": "millisecond"
+#                 },
+#                 {
+#                     "name": "c",
+#                     "type": "channel"
+#                 },
+#                 {
+#                     "name": "z",
+#                     "type": "space",
+#                     "unit": "micrometer"
+#                 },
+#                 {
+#                     "name": "y",
+#                     "type": "space",
+#                     "unit": "micrometer"
+#                 },
+#                 {
+#                     "name": "x",
+#                     "type": "space",
+#                     "unit": "micrometer"
+#                 }
+#             ],
+#             "datasets": datasets,
+#             "name": "/" + name,
+#             "translation": [
+#                 0,
+#                 0,
+#                 0.0,
+#                 0.0,
+#                 0.0
+#             ],
+#             "version": "0.4",
+#             "coordinateTransformations": [
+#                 {
+#                     "translation": [
+#                         0.0,
+#                         0.0,
+#                         position_xyz[2],
+#                         position_xyz[1],
+#                         position_xyz[0]
+#                     ],
+#                     "type": "translation"
+#                 }
+#             ]
+#         }
+#     ]
+#     attrs["omero"] = {
+#         "channels": [
+#             {
+#                 "active": True,
+#                 "coefficient": 1,
+#                 "color": "000000",
+#                 "family": "linear",
+#                 "inverted": False,
+#                 "label": "Channel:" + name + ":0",
+#                 "window": {
+#                     "end": 1.0,
+#                     "max": 1.0,
+#                     "min": 0.0,
+#                     "start": 0.0
+#                 }
+#             }
+#         ],
+#         "id": 1,
+#         "name": name,
+#         "rdefs": {
+#             "defaultT": 0,
+#             "defaultZ": 288,
+#             "model": "color"
+#         },
+#         "version": "0.4"
+#     }
+#     return attrs
 
 
 class TiffToNGFFException(Exception):
@@ -655,6 +655,8 @@ def write_ims_to_zarr(
     dataset = f['DataSet']['ResolutionLevel 0']['TimePoint 0']['Channel 0']['Data']
     
     joined_shapes = dataset.shape
+    print("ims_to_ngff dataset shape:" + str(joined_shapes))
+    return
 
     if deskew_options and deskew_options["deskew_method"] == "ps":
         block_size = chunk_size[2]
@@ -750,67 +752,67 @@ def ims_to_ngff_group(ims_fn, output, *args, **kwargs):
         print('unknown output format: ' + output)
 
 
-class DownsampleOptions(argschema.schemas.DefaultSchema):
-    block_divs = argschema.fields.List(
-        argschema.fields.Int, required=False, allow_none=True)
-    n_threads = argschema.fields.Int(required=False, allow_none=True)
+# class DownsampleOptions(argschema.schemas.DefaultSchema):
+#     block_divs = argschema.fields.List(
+#         argschema.fields.Int, required=False, allow_none=True)
+#     n_threads = argschema.fields.Int(required=False, allow_none=True)
 
 
-class DeskewOptions(argschema.schemas.DefaultSchema):
-    deskew_method = argschema.fields.Str(required=False, default='')
-    deskew_stride = argschema.fields.Int(required=False, default=None)
-    deskew_flip = argschema.fields.Bool(required=False, default=True)
-    deskew_crop = argschema.fields.Float(required=False, default=1.0)
+# class DeskewOptions(argschema.schemas.DefaultSchema):
+#     deskew_method = argschema.fields.Str(required=False, default='')
+#     deskew_stride = argschema.fields.Int(required=False, default=None)
+#     deskew_flip = argschema.fields.Bool(required=False, default=True)
+#     deskew_crop = argschema.fields.Float(required=False, default=1.0)
 
 
-class NGFFGenerationParameters(argschema.schemas.DefaultSchema):
-    output_format = argschema.fields.Str(required=True)
-    output_file = argschema.fields.Str(required=True)
-    max_mip = argschema.fields.Int(required=False, default=0)
-    concurrency = argschema.fields.Int(required=False, default=10)
-    compression = argschema.fields.Str(required=False, default="raw")
-    lvl_to_mip_kwargs = argschema.fields.Dict(
-        keys=argschema.fields.Int(),
-        values=argschema.fields.Nested(DownsampleOptions))
+# class NGFFGenerationParameters(argschema.schemas.DefaultSchema):
+#     output_format = argschema.fields.Str(required=True)
+#     output_file = argschema.fields.Str(required=True)
+#     max_mip = argschema.fields.Int(required=False, default=0)
+#     concurrency = argschema.fields.Int(required=False, default=10)
+#     compression = argschema.fields.Str(required=False, default="raw")
+#     lvl_to_mip_kwargs = argschema.fields.Dict(
+#         keys=argschema.fields.Int(),
+#         values=argschema.fields.Nested(DownsampleOptions))
 
-    # FIXME argschema supports lists and tuples,
-    #   but has some version differences
+#     # FIXME argschema supports lists and tuples,
+#     #   but has some version differences
 
-    mip_dsfactor = argschema.fields.Tuple((
-        argschema.fields.Int(),
-        argschema.fields.Int(),
-        argschema.fields.Int()), required=False, default=(2, 2, 2))
-    deskew_options = argschema.fields.Nested(
-        DeskewOptions, required=False)
-
-
-class NGFFGroupGenerationParameters(NGFFGenerationParameters):
-    group_names = argschema.fields.List(
-        argschema.fields.Str, required=True)
-    group_attributes = argschema.fields.List(
-        argschema.fields.Dict(required=False, default={}), default=[],
-        required=False)
+#     mip_dsfactor = argschema.fields.Tuple((
+#         argschema.fields.Int(),
+#         argschema.fields.Int(),
+#         argschema.fields.Int()), required=False, default=(2, 2, 2))
+#     deskew_options = argschema.fields.Nested(
+#         DeskewOptions, required=False)
 
 
-class IMSToNGFFParameters(NGFFGroupGenerationParameters):
-    input_file = argschema.fields.Str(required=True)
-    interleaved_channels = argschema.fields.Int(required=False, default=1)
-    channel = argschema.fields.Int(required=False, default=0)
+# class NGFFGroupGenerationParameters(NGFFGenerationParameters):
+#     group_names = argschema.fields.List(
+#         argschema.fields.Str, required=True)
+#     group_attributes = argschema.fields.List(
+#         argschema.fields.Dict(required=False, default={}), default=[],
+#         required=False)
 
 
-class IMSToZarrInputParameters(argschema.ArgSchema,
-                                   IMSToNGFFParameters):
-    chunk_size = argschema.fields.Tuple((
-        argschema.fields.Int(),
-        argschema.fields.Int(),
-        argschema.fields.Int(),
-        argschema.fields.Int(),
-        argschema.fields.Int()), required=False, default=(1, 1, 64, 64, 64))
+# class IMSToNGFFParameters(NGFFGroupGenerationParameters):
+#     input_file = argschema.fields.Str(required=True)
+#     interleaved_channels = argschema.fields.Int(required=False, default=1)
+#     channel = argschema.fields.Int(required=False, default=0)
+
+
+# class IMSToZarrInputParameters(argschema.ArgSchema,
+#                                    IMSToNGFFParameters):
+#     chunk_size = argschema.fields.Tuple((
+#         argschema.fields.Int(),
+#         argschema.fields.Int(),
+#         argschema.fields.Int(),
+#         argschema.fields.Int(),
+#         argschema.fields.Int()), required=False, default=(1, 1, 64, 64, 64))
 
 
 
 class IMSToZarr(argschema.ArgSchemaParser):
-    default_schema = IMSToZarrInputParameters
+    default_schema = TiffDirToZarrInputParameters #IMSToZarrInputParameters
 
     def run(self):
         deskew_options = (self.args["deskew_options"]
