+++ acpreprocessing/stitching_modules/convert_to_n5/tiff_to_ngff.py
@@ -19,7 +19,6 @@ import argschema
 import acpreprocessing.utils.convert
 import acpreprocessing.stitching_modules.convert_to_n5.psdeskew as psd
 
-
 def iterate_chunks(it, slice_length):
     """given an iterator, iterate over tuples of a
     given length from that iterator
@@ -97,14 +96,47 @@ def iterate_2d_arrays_from_mimgfns(mimgfns, interleaved_channels=1, channel=0):
             offset = (offset + r.get_length()) % interleaved_channels
 
 
-def iterate_numpy_chunks_from_mimgfns(
-        mimgfns, slice_length=None, pad=True, *args, **kwargs):
-    """iterate over a contiguous iterator of imageio multi-image files as
-    chunks of numpy arrays
+# def iterate_numpy_chunks_from_mimgfns(
+#         mimgfns, slice_length=None, pad=True, *args, **kwargs):
+#     """iterate over a contiguous iterator of imageio multi-image files as
+#     chunks of numpy arrays
+
+#     Parameters
+#     ----------
+#     mimgfns : list of str
+#         imageio-compatible name inputs to be opened as multi-images
+#     slice_length : int
+#         number of 2d arrays from mimgfns included per chunk
+#     pad : bool, optional
+#         whether to extend final returned chunks with zeros
+
+#     Yields
+#     ------
+#     arr : numpy.ndarray
+#         3D numpy array representing a consecutive chunk of 2D arrays
+#     """
+#     array_gen = iterate_2d_arrays_from_mimgfns(mimgfns, *args, **kwargs)
+#     for chunk in iterate_chunks(array_gen, slice_length):
+#         arr = numpy.array(chunk)
+#         if pad:
+#             if arr.shape[0] != slice_length:
+#                 newarr = numpy.zeros((slice_length, *arr.shape[1:]),
+#                                      dtype=arr.dtype)
+#                 newarr[:arr.shape[0], :, :] = arr[:, :, :]
+#                 yield newarr
+#             else:
+#                 yield arr
+#         else:
+#             yield arr
+
+
+def iterate_numpy_chunks_from_dataset(
+        dataset, slice_length=None, pad=True, *args, **kwargs):
+    """iterate over a contiguous hdf5 daataset as chunks of numpy arrays
 
     Parameters
     ----------
-    mimgfns : list of str
+    dataset : hdf5 dataset
         imageio-compatible name inputs to be opened as multi-images
     slice_length : int
         number of 2d arrays from mimgfns included per chunk
@@ -116,13 +148,13 @@ def iterate_numpy_chunks_from_mimgfns(
     arr : numpy.ndarray
         3D numpy array representing a consecutive chunk of 2D arrays
     """
-    array_gen = iterate_2d_arrays_from_mimgfns(mimgfns, *args, **kwargs)
-    for chunk in iterate_chunks(array_gen, slice_length):
-        arr = numpy.array(chunk)
+    #array_gen = iterate_2d_arrays_from_dataset(mimgfns, *args, **kwargs)
+    for chunk in iterate_chunks(dataset, slice_length,*args,**kwargs):
+        arr = numpy.asarray(chunk)
         if pad:
             if arr.shape[0] != slice_length:
                 newarr = numpy.zeros((slice_length, *arr.shape[1:]),
-                                     dtype=arr.dtype)
+                                      dtype=arr.dtype)
                 newarr[:arr.shape[0], :, :] = arr[:, :, :]
                 yield newarr
             else:
@@ -415,15 +447,125 @@ class MIPArray:
     end: int
 
 
-def iterate_mip_levels_from_mimgfns(
-        mimgfns, lvl, block_size, slice_length, downsample_factor,
+# def iterate_mip_levels_from_mimgfns(
+#         mimgfns, lvl, block_size, slice_length, downsample_factor,
+#         downsample_method=None, lvl_to_mip_kwargs=None,
+#         interleaved_channels=1, channel=0, deskew_kwargs={}):
+#     """recursively generate MIPmap levels from an iterator of multi-image files
+
+#     Parameters
+#     ----------
+#     mimgfns : list of str
+#         imageio-compatible name inputs to be opened as multi-images
+#     lvl : int
+#         integer mip level to generate
+#     block_size : int
+#         number of 2D arrays in chunk to process for a chunk at lvl
+#     slice_length : int
+#         number of 2D arrays to gather from tiff stacks
+#     downsample_factor : tuple of int
+#         integer downsampling factor for MIP levels
+#     downsample_method : str, optional
+#         downsampling method for
+#         acpreprocessing.utils.convert.downsample_stack_volume
+#     lvl_to_mip_kwargs :  dict, optional
+#         mapping of MIP level to kwargs used in MIPmap generation
+#     interleaved_channels : int
+#         number of channels interleaved in the tiff files (default 1)
+#     channel : int, optional
+#         channel from which interleaved data should be read (default 0)
+#     deskew_kwargs : dict, optional
+#         parameters for pixel shifting deskew
+
+#     Yields
+#     ------
+#     ma : acpreprocessing.stitching_modules.convert_to_n5.tiff_to_n5.MipArray
+#         object describing chunked array, MIP level of origin, and chunk indices
+#     """
+#     lvl_to_mip_kwargs = ({} if lvl_to_mip_kwargs is None
+#                          else lvl_to_mip_kwargs)
+#     mip_kwargs = lvl_to_mip_kwargs.get(lvl, {})
+#     start_index = 0
+#     chunk_index = 0
+#     if lvl > 0:
+#         num_chunks = downsample_factor[0]
+
+#         i = 0
+#         for ma in iterate_mip_levels_from_mimgfns(
+#                 mimgfns, lvl-1, block_size, slice_length,
+#                 downsample_factor, downsample_method,
+#                 lvl_to_mip_kwargs, interleaved_channels=interleaved_channels,
+#                 channel=channel, deskew_kwargs=deskew_kwargs):
+#             chunk = ma.array
+#             # throw array up for further processing
+#             yield ma
+#             # ignore if not parent resolution
+#             if ma.lvl != lvl-1:
+#                 continue
+
+#             try:
+#                 temp_lminus1_arr
+#             except NameError:
+#                 temp_lminus1_arr = numpy.empty(
+#                     (num_chunks*chunk.shape[0], *chunk.shape[1:]),
+#                     dtype=chunk.dtype)
+
+#             # fill in temporary block according to index
+#             chunk_size = chunk.shape[0]
+#             block_offset = i*block_size
+
+#             temp_lminus1_arr[
+#                 block_offset:block_offset+chunk_size, :, :] = chunk[:, :, :]
+
+#             # copy op only for uneven final chunk
+#             if chunk_size != block_size:
+#                 temp_lminus1_arr = temp_lminus1_arr[
+#                     :block_offset+chunk_size, :, :]
+
+#             if i == num_chunks - 1:
+#                 temp_arr = (
+#                     downsample_array(
+#                         temp_lminus1_arr, downsample_factor, dtype=chunk.dtype,
+#                         method=downsample_method, **mip_kwargs))
+#                 end_index = start_index + temp_arr.shape[0]
+#                 yield MIPArray(lvl, temp_arr, start_index, end_index)
+#                 start_index += temp_arr.shape[0]
+#                 i = 0
+#             else:
+#                 i += 1
+#         # add any leftovers
+#         if i > 0:
+#             temp_arr = acpreprocessing.utils.convert.downsample_stack_volume(
+#                 temp_lminus1_arr, downsample_factor, dtype=chunk.dtype,
+#                 method=downsample_method)
+#             end_index = start_index + temp_arr.shape[0]
+#             yield MIPArray(lvl, temp_arr, start_index, end_index)
+#     else:
+#         # get level 0 chunks
+#         # block_size is the number of slices to read from tiffs
+#         for chunk in iterate_numpy_chunks_from_mimgfns(
+#                 mimgfns, slice_length, pad=False,
+#                 interleaved_channels=interleaved_channels,
+#                 channel=channel):
+#             # deskew level 0 chunk
+#             if deskew_kwargs:
+#                 chunk = numpy.transpose(psd.deskew_block(
+#                     chunk, chunk_index, **deskew_kwargs), (2, 1, 0))
+#             end_index = start_index + chunk.shape[0]
+#             yield MIPArray(lvl, chunk, start_index, end_index)
+#             start_index += chunk.shape[0]
+#             chunk_index += 1
+
+
+def iterate_mip_levels_from_dataset(
+        dataset, lvl, block_size, slice_length, downsample_factor,
         downsample_method=None, lvl_to_mip_kwargs=None,
         interleaved_channels=1, channel=0, deskew_kwargs={}):
     """recursively generate MIPmap levels from an iterator of multi-image files
 
     Parameters
     ----------
-    mimgfns : list of str
+    dataset : hdf5 dataset
         imageio-compatible name inputs to be opened as multi-images
     lvl : int
         integer mip level to generate
@@ -451,7 +593,7 @@ def iterate_mip_levels_from_mimgfns(
         object describing chunked array, MIP level of origin, and chunk indices
     """
     lvl_to_mip_kwargs = ({} if lvl_to_mip_kwargs is None
-                         else lvl_to_mip_kwargs)
+                          else lvl_to_mip_kwargs)
     mip_kwargs = lvl_to_mip_kwargs.get(lvl, {})
     start_index = 0
     chunk_index = 0
@@ -459,8 +601,8 @@ def iterate_mip_levels_from_mimgfns(
         num_chunks = downsample_factor[0]
 
         i = 0
-        for ma in iterate_mip_levels_from_mimgfns(
-                mimgfns, lvl-1, block_size, slice_length,
+        for ma in iterate_mip_levels_from_dataset(
+                dataset, lvl-1, block_size, slice_length,
                 downsample_factor, downsample_method,
                 lvl_to_mip_kwargs, interleaved_channels=interleaved_channels,
                 channel=channel, deskew_kwargs=deskew_kwargs):
@@ -511,8 +653,8 @@ def iterate_mip_levels_from_mimgfns(
     else:
         # get level 0 chunks
         # block_size is the number of slices to read from tiffs
-        for chunk in iterate_numpy_chunks_from_mimgfns(
-                mimgfns, slice_length, pad=False,
+        for chunk in iterate_numpy_chunks_from_dataset(
+                dataset, slice_length, pad=False,
                 interleaved_channels=interleaved_channels,
                 channel=channel):
             # deskew level 0 chunk
@@ -570,7 +712,8 @@ def write_mimgfns_to_n5(
     """
     group_attributes = ([] if group_attributes is None else group_attributes)
     deskew_options = ({} if deskew_options is None else deskew_options)
-
+    
+    array_gen = iterate_2d_arrays_from_mimgfns(mimgfns, interleaved_channels=interleaved_channels,channel=channel)
     joined_shapes = joined_mimg_shape_from_fns(
         mimgfns, concurrency=concurrency,
         interleaved_channels=interleaved_channels, channel=channel)
@@ -632,8 +775,8 @@ def write_mimgfns_to_n5(
 
         with concurrent.futures.ThreadPoolExecutor(max_workers=workers) as e:
             futs = []
-            for miparr in iterate_mip_levels_from_mimgfns(
-                    mimgfns, max_mip, block_size, slice_length, mip_dsfactor,
+            for miparr in iterate_mip_levels_from_dataset(
+                    array_gen, max_mip, block_size, slice_length, mip_dsfactor,
                     lvl_to_mip_kwargs=lvl_to_mip_kwargs,
                     interleaved_channels=interleaved_channels,
                     channel=channel, deskew_kwargs=deskew_kwargs):
@@ -798,7 +941,8 @@ def write_mimgfns_to_zarr(
     """
     group_attributes = ([] if group_attributes is None else group_attributes)
     deskew_options = ({} if deskew_options is None else deskew_options)
-
+    
+    array_gen = iterate_2d_arrays_from_mimgfns(mimgfns, interleaved_channels=interleaved_channels,channel=channel)
     joined_shapes = joined_mimg_shape_from_fns(
         mimgfns, concurrency=concurrency,
         interleaved_channels=interleaved_channels, channel=channel)
@@ -863,8 +1007,8 @@ def write_mimgfns_to_zarr(
 
         with concurrent.futures.ThreadPoolExecutor(max_workers=workers) as e:
             futs = []
-            for miparr in iterate_mip_levels_from_mimgfns(
-                    mimgfns, max_mip, block_size, slice_length, mip_dsfactor,
+            for miparr in iterate_mip_levels_from_dataset(
+                    array_gen, max_mip, block_size, slice_length, mip_dsfactor,
                     lvl_to_mip_kwargs=lvl_to_mip_kwargs,
                     interleaved_channels=interleaved_channels,
                     channel=channel, deskew_kwargs=deskew_kwargs):
@@ -905,6 +1049,7 @@ class DeskewOptions(argschema.schemas.DefaultSchema):
     deskew_method = argschema.fields.Str(required=False, default='')
     deskew_stride = argschema.fields.Int(required=False, default=None)
     deskew_flip = argschema.fields.Bool(required=False, default=True)
+    deskew_transpose = argschema.fields.Bool(required=False, default=False)
     deskew_crop = argschema.fields.Float(required=False, default=1.0)
